file_list: implement file_list_add_cdev_entry
Follow-up commit will add a first user that wants to add a cdev
and more may follow, so let's provide a common
file_list_add_cdev_entry they can call.

The benefit of doing that in common/file-list.c is that we can avoid
an extra allocation for prefixing /dev.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20230609075229.1093397-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>