scripts: add container.sh as simple podman wrapper
We have a Containerfile in tree now that is used for CI, but it's not
completely self evident, how to use it locally. Add a simple script
that mounts the current working directory into the container and calls
its arguments inside.

Example building MAKEALL (Some configs will expectedly fail due to size
or firmware constraints):

  ./scripts/container.sh ./MAKEALL

Example starting a shell:

  ./scripts/container.sh /bin/bash

Signed-off-by: Ahmad Fatoum <ahmad@a3f.at>
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>