test/Containerfile: run wget with no-verbose flag
Container build results in a number of toolchain downloads and default
wget behavior of updating the progress bar results in a lot of lines
to wade through when reviewing the log. There's middle ground between
this and quiet operation in the form of the --no-verbose or -nv flag.
Make use of it.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20230612124728.1080623-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>