scripts: remove openssl dependency in multi_v7_defconfig
The 32-bit rockchip support doesn't use the rkimage tool and instead
depends on mkimage to wrap the barebox binary. There's thus no need to
compile the rkimage utility for 32-bit barebox builds. This is a good
thing, because it removes the openssl dependency as the library is used
for sha256/sha512 calculation.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20230612125331.1085059-2-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>