test: self: printf: silence clang warnings
We added -Wno-format-security a short while back, because we didn't want
GCC to warn us about possible format string issues in the self test that's
meant to exercise the printf facility. In a similar vein, clang also
sees potential problems in the code, so hide it, just for that file,
with -Wno-format.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20230612130239.1087599-9-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>