crypto: caam - fix pointer to u32 casts
For 64-bit systems, we only have CAAM initialization support in PBL
running as BL2 in EL3, but nothing for barebox proper. Understandably,
trying to compile the CAAM code for 64-bit leads to warnings about casts
between pointers and smaller integers. A sync with Linux to get 64-bit
support is probably in-order for barebox proper CAAM drivers to actually
work on i.MX8M, but for now, let's just fix the warnings.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20230612130239.1087599-10-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>