mci: sdhci: dove: fix pointer to u32 casts
The dove SDHCI driver doesn't fully work on 64-bit, which
is ok as the SoC is ARMv7-only anways. We do want to be able to build
test the driver though, so replace straight casts of addr with
lower_32_bits(virt_to_phys(addr)), which has the same result, but makes
it clear that the upper 32 bit are truncated and silences the compiler
warnings.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20230612130239.1087599-12-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>