boot: don't print error message on -ENOMEDIUM
It can be expected for a boot entry to usually fail, e.g. when it checks
for a removable USB drive that's not always connected. Such boot targets
have the choice of either returning 0, which means it was a dry run and
boot aborts or an error code, which yields an error message.

Let's handle -ENOMEDIUM specially and not print an error for it, so it
can be used in such scenarios.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20230612130322.1093717-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>