ci: build: have CI build ARCH=openrisc
qemu-system-misc ships with qemu-system-or1k and kernel.org mirrors a
crostool toolchain capable of building barebox for running in it.
Let's make use of that. We don't yet activate openrisc for CI test
as it fails to boot up with SELFTEST_OF_MANIPULATION. A suggested fix
is still under discussion[1].

[1]: https://lore.barebox.org/barebox/20230605063719.1431079-1-a.fatoum@pengutronix.de/

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20230619094637.4134449-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>