usb: add basic USB Type-C framework
Board code or user scripts may want to base their decisions on the
state of a USB-C connector:

  - Is the USB role that of a USB device or of a USB host?
  - Is the board being powered externally?
  - Is a debug accessory attached?

For this reason, we add here a very simplified Type-C driver core inspired
by Linux that just registers a device and populates it with three parameters:

  - $dev.usb_role = { device, host }
  - $dev.pwr_role = { sink, source }
  - $dev.accessory = { none, audio, debug }

These variables are updated by struct typec_operations::poll, which
corresponds to the IRQ handler for the Linux drivers. The poll function
is called on every variable access and early exits if no IRQ was
indicated.

In future, we may elect to register the poll function with the poller
framework, so it's called cyclically, like we do for e.g. Ethernet link
detection, but for now, this is probably sufficient.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20230621090329.2652672-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>