soc: imx: have IMX8M_FEATCTRL select IMX_OCOTP
IMX8M_FEATCTRL selects FEATURE_CONTROLLER, so of_feature_controller_check()
will return an error if the feature controller driver is missing.

Given that the default for IMX_OCOTP is n, but for IMX8M_FEATCTRL is y,
this error case is easily hit. Resolve this by selecting IMX_OCOTP when
IMX8M_FEATCTRL is enabled.

Reported-by: Bastian Krause <bst@pengutronix.de>
Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20230703172054.536931-2-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>