mci: imx-esdhc-pbl: fix image load in DDR mode
The __esdhc_send_cmd checks the sdhci.timing setting and configures it
accordingly. If the BootROM configured the device to operare in DDR mode
we need to honor that else the PBL can't load the binary correctly from
the eMMC. Therefore readback the sdhci mixctrl setting and set the
sdhci.timing to DDR52. At the moment DDR52 is the fastest/highest
transferrate the BootROM supports, so we don't need to handle HS200/400
yet.

Signed-off-by: Marco Felsch <m.felsch@pengutronix.de>
Reviewed-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20230705190352.1448472-1-m.felsch@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>