MIPS: addrspace: rectify ksseg segment naming
There was a slight confusion in macro definitions for various memory
segments. In the 64-bit mode, there was CKSEG2ADDR defined without a
corresponding CKSEG2 definition, which led to errors when trying to
actually use it. Instead there was a CKSSEG definition that stands for
exactly the same memory segment.

Change the name of this memory segment macro into KSSEG/CKSSEG for the
32/64-bit mode respectively and also rename the corresponding *ADDR
macros as this is how this segment is actually named in the MIPS
architecture specification. This makes the naming more consistent.

Signed-off-by: Denis Orlov <denorl2009@gmail.com>
Link: https://lore.barebox.org/20230725050618.3451-2-denorl2009@gmail.com
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>