MIPS: pbl: make sure to disable interrupts/watchpoints on entry
The first thing that we should do on entry is to disable interrupts and
watchpoints, as who knows in what state we got there. The former was not
being done consistently on every board, while the latter was only being
done in the barebox proper for some reason. Add a macro that combines
both of those actions and ensure that it's used in every board's entry
code.

Signed-off-by: Denis Orlov <denorl2009@gmail.com>
Link: https://lore.barebox.org/20230725050618.3451-13-denorl2009@gmail.com
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>