MIPS: clean up barebox proper entry point
For MIPS, we don't actually have boards which lack PBL in the resulting
image. As such, the code in _start() is mostly redundant: the stack was
setup in ENTRY_FUNCTION_END, while copying the code to the link location
is not needed - it should already be executing at TEXT_BASE, where we
have decompressed it. Other than that, disabling interrupts with
watchpoints and enabling 64-bit addressing was already moved into PBL,
so we don't really need to repeat it there.

Replace the assembly implementation of _start() with the C one, as the
only thing it now does is jump straight into 'relocate_code()'.

Signed-off-by: Denis Orlov <denorl2009@gmail.com>
Link: https://lore.barebox.org/20230725050618.3451-15-denorl2009@gmail.com
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>