commands: go: support passing integer arguments instead of argc/argv
The go command accepts extra arguments, which are passed through to the
application using two parameters: argc to hold the number of arguments
and argv as a NULL pointer terminated array of pointers to
nul-terminated strings.

As the go command is meant to invoke bare metal applications, it is
useful to be able to pass arguments parsed to integers using the
default calling convention, so add a new -i argument. The old behavior
remains the default and a new -s arguments makes it explicit for
symmetry.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20230726125222.3706933-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>