Drop mkimage tool
The barebox sourcetree contains a very outdated U-Boot mkimage tool.
The only purpose of this tool is to create a uImage barebox image which
can be used to start barebox 2nd stage from U-Boot. There is also a Make
target for generating such an image. This however predates PBL support
and is incompatible with PBL enabled barebox builds.

As the barebox.uimage target is not very useful nowadays remove it.

A recent mkimage is availabe in most Linux distributions as a host tool,
so remove the outdated version from barebox as well.

Link: https://lore.barebox.org/20230803065749.3827198-1-s.hauer@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>