of: never fixup internal live tree
Calling of_fix_tree() on the internal live tree has undesired side
effects, so refrain from doing so.

One of the side effects is that some parts of barebox store pointers
to properties in the live tree which might become invalid when during
of_fixup these properties are deleted or updated.

The other is that it's unexpected that the live tree is modified
after a dry run boot with the live tree.

This changes of_get_fixed_tree() to always call of_fix_tree() on
a copy of the device tree and not on the device tree itself. To
emphasize this make the device tree argument to of_get_fixed_tree()
const.

Reviewed-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20230803085134.3912884-1-s.hauer@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>