boot: support running scripts when doing a dry run
Dry runs are quite useful for debugging or for time measurement, but so
far they weren't useful with scripts as they just did a very early exit.

This is understandable as scripts don't necessarily observe the
convention that everything done before the dry run should not have an
impact on later boot runs. For the cases, where the bootscript is vetted
to behave properly under a dry run, support specifing dryrun twice, so
the script is executed, but the automatic bootm afterwards isn't.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20230803135217.1770111-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>