console: allocate only once instead of twice per log message
We do two allocations for each log message, one for the log message
string itself and another for the struct log_entry referencing it.

We could just make the log message directly follow the log entry and
save the space for the pointer and the time to do a second allocation.

No functional change.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20230804081147.559379-2-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>