Convert License Identifiers to SPDX 3.0 for files originating from LiMon
The respective U-Boot commit was ported to Barebox in commit
0b5d36d77b6022bd1f40 (2023-07-27, Sascha Hauer: "Add SPDX License
identifier for files originating from LiMon"), but the U-Boot change was
already made in 2014, and the SPDX specification was updated since then
and deprecated the "GPL-2.0" identifier in favour of "GPL-2.0-only".
Update the license identifiers to SPDX 3.0 specification, which we also
use everywhere else in the Barebox code.

Signed-off-by: Roland Hieber <rhi@pengutronix.de>
Link: https://lore.barebox.org/20230804094650.2094025-1-rhi@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>