ARM: i.MX25: image: consult pkg-config for openssl cflags
For i.MX HABv3, we already ask pkg-config about the flags
to link the image generation utility against OpenSSL, but we assumed
that the headers would be found in the toolchain's default system header
search path. This may not always be the case, so explicitly ask
pkg-config about the CFLAGS as well.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20230807074445.1042385-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>