ARM: add Skov i.MX8MP based board support
Add basic support for the Skov i.MX8MP based system.
This is initial support and will be extended later at least with board version
detection.

The compatible used in the board file is for barebox specific generic
devicetree which is expected to work with all Skov i.MX8MP based boards.

Signed-off-by: Oleksij Rempel <o.rempel@pengutronix.de>
Reviewed-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20230814121234.2405038-1-o.rempel@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>