filetype: return error and pass filetype as pointer argument
file_name_detect_type(), file_name_detect_type_offset() and
cdev_detect_type() return the filetype. With this all errors from these
functions remain undetected and are just returned as filetype_unknown.

Explicitly return an error code and pass the filetype as pointer
argument so that callers can detect and handle errors.

This fixes a bug in the bootm code where the returned filetype was
erroneously tested for being smaller than 0. This was never true
and so the corresponding error message was never printed. Now with
this patch a non existing initrd or device tree file is responded
with a meaningful error message.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>