ARM: boards: polyhex-debix: fix RGMII pin setup
On the Debix SBC the EQOS interface is used as primary interface, the
FEC pins are routed to the extension header. On the other hand the Debix
SoM/Baseboard combination both interfaces are used. So fix the RGMII pin
setup by adding the EQOS pin as well and keep the FEC.

Signed-off-by: Marco Felsch <m.felsch@pengutronix.de>
Link: https://lore.barebox.org/20230829124356.4178308-1-m.felsch@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>